apps/ai-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
redis[hiredis]==5.0.0
qdrant-client==1.11.0
sqlalchemy[asyncio]==2.0.32
pymssql==2.3.0
google-generativeai==0.8.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
PyPDF2==3.0.1
tiktoken==0.7.0
httpx==0.27.0
python-multipart==0.0.9
apscheduler==3.10.4
